requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==20.1.0
bcrypt==3.2.0
cached-property==1.5.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
distro==1.5.0
Django==3.1
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.11.1
docker==4.3.1
docker-compose==1.26.2
dockerpty==0.4.1
docopt==0.6.2
flake8==3.8.3
flake8-django==1.1.1
gunicorn==20.0.4
idna==2.10
isort==5.4.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
paramiko==2.7.1
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
PyNaCl==1.4.0
pyrsistent==0.16.0
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
screen==1.0.1
six==1.15.0
sqlparse==0.3.1
texttable==1.6.2
toml==0.10.1
urllib3==1.25.10
websocket-client==0.57.0
wrapt==1.12.1
xlwt==1.3.0
